added internal app record support in sessionbear
