+++ browser-extension/sessionbear/rollup.config.js
@@ -133,9 +133,9 @@ export default [
     plugins: commonPlugins,
   },
   {
-    input: "src/client-scripts/index.ts",
+    input: "src/page-scripts/sessionRecorderHelper.js",
     output: {
-      file: `${OUTPUT_DIR}/client.js`,
+      file: `${OUTPUT_DIR}/page-scripts/sessionRecorderHelper.ps.js`,
       format: "iife",
     },
     plugins: commonPlugins,
