--- browser-extension/sessionbear/src/client-scripts/index.ts
@@ -1,1 +1,0 @@
-import "./sessionRecorderHelper";
