+++ browser-extension/sessionbear/src/content-scripts/app/index.ts
@@ -1,4 +1,8 @@
-import { initMessageHandler, initExtensionMessageListener } from "./messageHandler";
+import { EXTENSION_MESSAGES } from "../../constants";
+import { isExtensionEnabled } from "../../utils";
+import { initExtensionMessageListener } from "../common/extensionMessageListener";
+import { initSessionRecording } from "../common/sessionRecorder";
+import { initMessageHandler } from "./messageHandler";
 
 //document.documentElement.setAttribute("rq-ext-version", chrome.runtime.getManifest()["version"]);
 
@@ -10,3 +14,9 @@ import { initMessageHandler, initExtensionMessageListener } from "./messageHandl
 
 initMessageHandler();
 initExtensionMessageListener();
+isExtensionEnabled().then((isExtensionStatusEnabled) => {
+  if (isExtensionStatusEnabled) {
+    chrome.runtime.sendMessage({ action: EXTENSION_MESSAGES.HANDSHAKE_CLIENT });
+    initSessionRecording();
+  }
+});
