+++ browser-extension/sessionbear/src/page-scripts/sessionRecorderHelper.js
