+++ browser-extension/sessionbear/src/service-worker/services/clientHandler.ts
@@ -1,27 +1,26 @@
-import { generateUrlPattern } from "../../utils";
-import { WEB_URL, OTHER_WEB_URLS } from "../../../../config/dist/config.build.json";
-import { isExtensionEnabled } from "./utils";
+// import { generateUrlPattern } from "../../utils";
+// import { WEB_URL, OTHER_WEB_URLS } from "../../../../config/dist/config.build.json";
+import { isExtensionEnabled } from "../../utils";
 import { Variable, onVariableChange } from "../variable";
 
-const CLIENT_SCRIPT_ID = "client-script";
-const excludeMatchesPatterns = [WEB_URL, ...OTHER_WEB_URLS].map(generateUrlPattern).filter((pattern) => !!pattern);
+// const excludeMatchesPatterns = [WEB_URL, ...OTHER_WEB_URLS].map(generateUrlPattern).filter((pattern) => !!pattern);
+
+const CLIENT_SCRIPTS: chrome.scripting.RegisteredContentScript[] = [
+  {
+    id: "page-script-sessionRecorder",
+    js: ["page-scripts/sessionRecorderHelper.ps.js"],
+    world: "MAIN",
+    persistAcrossSessions: false,
+    matches: ["http://*/*", "https://*/*"],
+    runAt: "document_start",
+  },
+];
 
 /** Loading AJAX Interceptor ASAP */
-const registerClientScript = async () => {
+const registerClientScripts = async () => {
   console.log("[registerClientScript]");
   chrome.scripting
-    .registerContentScripts([
-      {
-        id: CLIENT_SCRIPT_ID,
-        js: ["client.js"],
-        world: "MAIN",
-        allFrames: true,
-        persistAcrossSessions: false,
-        matches: ["http://*/*", "https://*/*"],
-        runAt: "document_start",
-        excludeMatches: excludeMatchesPatterns,
-      },
-    ])
+    .registerContentScripts(CLIENT_SCRIPTS)
     .then(() => {
       console.log("[registerClientScript]");
       chrome.scripting
@@ -31,10 +30,10 @@ const registerClientScript = async () => {
     .catch((err) => console.warn("[unregisterClientScript]", "unexpected error", err));
 };
 
-const unregisterClientScript = async () => {
+const unregisterClientScripts = async () => {
   console.log("[unregisterClientScript]");
   return chrome.scripting
-    .unregisterContentScripts({ ids: [CLIENT_SCRIPT_ID] })
+    .unregisterContentScripts({ ids: CLIENT_SCRIPTS.map((script) => script.id) })
     .then(() => {
       console.log("[unregisterClientScript]", "unregisterClientScript complete");
       chrome.scripting
@@ -47,9 +46,9 @@ const unregisterClientScript = async () => {
 const setupClientScript = async (isExtensionStatusEnabled: boolean) => {
   console.log("[initClientHandler.setupClientScript]", { isExtensionEnabled });
   if (isExtensionStatusEnabled) {
-    registerClientScript();
+    registerClientScripts();
   } else {
-    unregisterClientScript();
+    unregisterClientScripts();
   }
 };
