+++ browser-extension/sessionbear/src/service-worker/services/contextMenu.ts
@@ -1,6 +1,6 @@
+import { isExtensionEnabled } from "../../utils";
 import { onVariableChange, setVariable, Variable } from "../variable";
 import extensionIconManager from "./extensionIconManager";
-import { isExtensionEnabled } from "./utils";
 
 // TODO: fix circular dependency
 // import { sendMessageToApp } from "./messageHandler";
